Remove audio stall/lag by never waiting for blocks
First moved all api calls onto another thread, but that was a
small win (1-3 ms).  The real problem was a bug where the
queue would wait for a free block even when the read (exactly)
fit the size of the remaining space in the current block.

Now that check/wait is made at the beginning and avoided if
a block is already available, and so by checking the block
availability before calling write, you can avoid the wait
altogether. Which is what was thought to be happening, but
the bug prevented.